Make BugDropStart default to FORCE_OFF, not AUTO.
Since it's a manually-enabled bug compatibility mode, AUTO isn't one of
the available UI options.

This was causing Windows PuTTY to display a blank entry in the drop-down
for "Discards data sent before its greeting".

(It is possible that this unhelpful default has escaped into saved
sessions of snapshot users, which would have the same effect, but since
the actual using code can cope with it, I've not done anything to clean
that up.)