Fix obvious braino in tempseat_output.
I'd added a data length to the 'type' field of output_chunk instead of
the 'size' field.

Caused an assertion failure when I tried a simple SSH proxy operation
on Windows just now, because the output_chunks and the output bufchain
didn't match up, and no wonder. The mystery is how it hasn't been
consistently failing like that since September!