testcrypt: move enumerated types out into another header.
Now all the lookup functions like get_hashalg() and
get_primegenpolicy() are autogenerated by macro expansion from a
header a little like the existing testcrypt.h. But the macros are much
simpler this time.

This doesn't really change anything, _except_ that now I'll be able to
reinclude the same header to do other things with the list of
enumerated types.