Factor out term_request_resize().
This tiny refactoring makes a convenient function for setting all the
'pending' flags and triggering a callback for the next window update.

This saves a bit of code, but that's not really the main point (or
else I'd have done the same to all the other similar things like
window moves). The point is that in a future commit I'm going to want
to do an extra thing on every server-controlled window resize, and
this refactoring gives me a single place to put that extra action.